requirements.txt
astroid==1.6.6
backports.functools-lru-cache==1.6.1
certifi==2020.6.20
chardet==3.0.4
configparser==4.0.2
enum34==1.1.10
futures==3.3.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.5.1
mccabe==0.6.1
pylint==1.9.5
requests==2.24.0
singledispatch==3.4.0.3
six==1.15.0
urllib3==1.25.10
virtualenv==16.7.9
wrapt==1.12.1
django==3.1.1
gunicorn==19.7.1
whitenoise==3.3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
pytz==2020.1